fix: Correcciones tras la revision refs #170
