+++ src/main/java/acme/constraints/AssistanceAgentValidator.java
@@ -5,7 +5,7 @@ import javax.validation.ConstraintValidatorContext;
 
 import acme.client.components.validation.AbstractValidator;
 import acme.client.components.validation.Validator;
-import acme.entities.assistance_agents.AssistanceAgent;
+import acme.realms.assistance_agents.AssistanceAgent;
 
 @Validator
 public class AssistanceAgentValidator extends AbstractValidator<ValidAssistanceAgent, AssistanceAgent> {
@@ -24,23 +24,26 @@ public class AssistanceAgentValidator extends AbstractValidator<ValidAssistanceA
 		if (assistanceAgent == null)
 			super.state(context, false, "*", "javax.validation.constraints.NotNull.message");
 		else {
-			boolean codeContaintsInitials;
+			boolean codeContainsInitials = true;
 
 			try {
 				String code = assistanceAgent.getEmployeeCode();
 				String name = assistanceAgent.getUserAccount().getIdentity().getName();
 				String surname = assistanceAgent.getUserAccount().getIdentity().getSurname();
 
-				String initials = code.substring(0, 2);
+				char codeFirstChar = Character.toUpperCase(code.charAt(0));
+				char codeSecondChar = Character.toUpperCase(code.charAt(1));
+				char nameFirstChar = Character.toUpperCase(name.charAt(0));
+				char surnameFirstChar = Character.toUpperCase(surname.charAt(0));
 
-				String expectedInitials = (name.substring(0, 1) + surname.substring(0, 1)).toUpperCase();
-				codeContaintsInitials = initials == expectedInitials;
+				if (!(codeFirstChar == nameFirstChar && codeSecondChar == surnameFirstChar))
+					codeContainsInitials = false;
 
 			} catch (Error e) {
-				codeContaintsInitials = false;
+				codeContainsInitials = false;
 			}
 
-			super.state(context, codeContaintsInitials, "*", "acme.validation.role.identifier.message");
+			super.state(context, codeContainsInitials, "*", "acme.validation.role.identifier.message");
 		}
 
 		result = !super.hasErrors(context);
