+++ src/main/java/acme/constraints/CustomerValidator.java
@@ -5,7 +5,7 @@ import javax.validation.ConstraintValidatorContext;
 
 import acme.client.components.validation.AbstractValidator;
 import acme.client.components.validation.Validator;
-import acme.entities.customers.Customer;
+import acme.realms.customer.Customer;
 
 @Validator
 public class CustomerValidator extends AbstractValidator<ValidCustomer, Customer> {
@@ -24,32 +24,26 @@ public class CustomerValidator extends AbstractValidator<ValidCustomer, Customer
 		if (customer == null)
 			super.state(context, false, "*", "javax.validation.constraints.NotNull.message");
 		else {
-			boolean codeContaintsInitials;
+			boolean codeContainsInitials = true;
 
 			try {
 				String code = customer.getIdentifier();
 				String name = customer.getUserAccount().getIdentity().getName();
 				String surname = customer.getUserAccount().getIdentity().getSurname();
 
-				int length = code.length();
-				String initials;
-				if (length == 8)
-					initials = code.substring(0, 2);
-				else
-					initials = code.substring(0, 3);
+				char codeFirstChar = Character.toUpperCase(code.charAt(0));
+				char codeSecondChar = Character.toUpperCase(code.charAt(1));
+				char nameFirstChar = Character.toUpperCase(name.charAt(0));
+				char surnameFirstChar = Character.toUpperCase(surname.charAt(0));
 
-				String expectedInitials = (name.substring(0, 1) + surname.substring(0, 1)).toUpperCase();
-				String[] surnameParts = surname.split("\\s+");
-				if (surnameParts.length > 1)
-					expectedInitials += surnameParts[1].substring(0, 1).toUpperCase();
-
-				codeContaintsInitials = initials == expectedInitials;
+				if (!(codeFirstChar == nameFirstChar && codeSecondChar == surnameFirstChar))
+					codeContainsInitials = false;
 
 			} catch (Error e) {
-				codeContaintsInitials = false;
+				codeContainsInitials = false;
 			}
 
-			super.state(context, codeContaintsInitials, "*", "acme.validation.customer.identifier.message");
+			super.state(context, codeContainsInitials, "*", "acme.validation.customer.identifier.message");
 		}
 
 		result = !super.hasErrors(context);
