+++ src/main/java/acme/entities/flight_assignments/FlightAssignment.java
@@ -15,8 +15,8 @@ import acme.client.components.validation.Mandatory;
 import acme.client.components.validation.Optional;
 import acme.client.components.validation.ValidMoment;
 import acme.client.components.validation.ValidString;
-import acme.entities.flight_crew_members.FlightCrewMember;
 import acme.entities.legs.Leg;
+import acme.realms.flight_crew_members.FlightCrewMember;
 import lombok.Getter;
 import lombok.Setter;
