+++ src/main/java/acme/entities/reviews/Review.java
@@ -46,7 +46,7 @@ public class Review extends AbstractEntity {
 	private String				text;
 
 	@Optional
-	@ValidNumber(min = 0, max = 10, integer = 2)
+	@ValidNumber(min = 0, max = 10)
 	@Automapped
 	private Double				score;
