+++ src/main/java/acme/entities/service/Service.java
@@ -3,8 +3,6 @@ package acme.entities.service;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.ManyToOne;
-import javax.validation.Valid;
 
 import acme.client.components.basis.AbstractEntity;
 import acme.client.components.datatypes.Money;
@@ -15,7 +13,6 @@ import acme.client.components.validation.ValidMoney;
 import acme.client.components.validation.ValidNumber;
 import acme.client.components.validation.ValidString;
 import acme.client.components.validation.ValidUrl;
-import acme.entities.airport.Airport;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -37,13 +34,13 @@ public class Service extends AbstractEntity {
 	private String				pictureUrl;
 
 	@Mandatory
-	@ValidNumber
+	@ValidNumber(min = 0, max = 100)
 	@Automapped
 	private Double				averageDwellTime;
 
 	@Optional
 	@ValidString(pattern = "^[A-Z]{4}-[0-9]{2}$")
-	@Automapped
+	@Column(unique = true)
 	private String				promotionCode;
 
 	@Optional
@@ -51,8 +48,4 @@ public class Service extends AbstractEntity {
 	@Column(unique = true)
 	private Money				discount;
 
-	@Mandatory
-	@Valid
-	@ManyToOne(optional = false)
-	private Airport				airport;
 }
