+++ src/main/java/acme/realms/airline_managers/AirlineManager.java
@@ -1,5 +1,5 @@
 
-package acme.entities.assistance_agents;
+package acme.realms.airline_managers;
 
 import java.util.Date;
 
@@ -11,15 +11,13 @@ import javax.persistence.TemporalType;
 import javax.validation.Valid;
 
 import acme.client.components.basis.AbstractRole;
-import acme.client.components.datatypes.Money;
 import acme.client.components.mappings.Automapped;
 import acme.client.components.validation.Mandatory;
 import acme.client.components.validation.Optional;
 import acme.client.components.validation.ValidMoment;
-import acme.client.components.validation.ValidMoney;
-import acme.client.components.validation.ValidString;
+import acme.client.components.validation.ValidNumber;
 import acme.client.components.validation.ValidUrl;
-import acme.constraints.ValidAssistanceAgent;
+import acme.constraints.ValidAirlineManager;
 import acme.constraints.ValidRoleIdentifier;
 import acme.entities.airlines.Airline;
 import lombok.Getter;
@@ -28,44 +26,34 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-@ValidAssistanceAgent
-public class AssistanceAgent extends AbstractRole {
+@ValidAirlineManager
+public class AirlineManager extends AbstractRole {
 
 	private static final long	serialVersionUID	= 1L;
 
 	@Mandatory
-	@Column(unique = true)
 	@ValidRoleIdentifier
-	private String				employeeCode;
+	@Column(unique = true)
+	private String				identifierNumber;
 
 	@Mandatory
-	@ValidString(max = 255)
+	@ValidNumber(min = 0)
 	@Automapped
-	private String				spokenLanguages;
+	private Integer				experience;
 
 	@Mandatory
 	@ValidMoment(past = true)
 	@Temporal(TemporalType.TIMESTAMP)
-	private Date				moment;
-
-	@Optional
-	@ValidString(max = 255)
-	@Automapped
-	private String				bio;
-
-	@Optional
-	@ValidMoney
-	@Automapped
-	private Money				salary;
-
-	@Optional
-	@ValidUrl
-	@Automapped
-	private String				photoUrl;
+	private Date				birthdate;
 
 	@Mandatory
 	@Valid
 	@ManyToOne(optional = false)
 	private Airline				airline;
+
+	@Optional
+	@ValidUrl
+	@Automapped
+	private String				linkPicture;
 
 }
