+++ src/main/java/acme/realms/assistance_agents/AssistanceAgent.java
@@ -1,9 +1,13 @@
 
-package acme.entities.flight_crew_members;
+package acme.realms.assistance_agents;
+
+import java.util.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
 import javax.validation.Valid;
 
 import acme.client.components.basis.AbstractRole;
@@ -11,13 +15,12 @@ import acme.client.components.datatypes.Money;
 import acme.client.components.mappings.Automapped;
 import acme.client.components.validation.Mandatory;
 import acme.client.components.validation.Optional;
+import acme.client.components.validation.ValidMoment;
 import acme.client.components.validation.ValidMoney;
-import acme.client.components.validation.ValidNumber;
 import acme.client.components.validation.ValidString;
-import acme.constraints.ValidFlightCrewMember;
-import acme.constraints.ValidPhoneNumber;
+import acme.client.components.validation.ValidUrl;
+import acme.constraints.ValidAssistanceAgent;
 import acme.constraints.ValidRoleIdentifier;
-
 import acme.entities.airlines.Airline;
 import lombok.Getter;
 import lombok.Setter;
@@ -25,43 +28,44 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-@ValidFlightCrewMember
-public class FlightCrewMember extends AbstractRole {
+@ValidAssistanceAgent
+public class AssistanceAgent extends AbstractRole {
 
 	private static final long	serialVersionUID	= 1L;
 
 	@Mandatory
-	@ValidRoleIdentifier
 	@Column(unique = true)
+	@ValidRoleIdentifier
 	private String				employeeCode;
 
 	@Mandatory
-	@ValidPhoneNumber
-	@Column(unique = true)
-	private String				phoneNumber;
+	@ValidString(max = 255)
+	@Automapped
+	private String				spokenLanguages;
 
 	@Mandatory
-	@ValidString
+	@ValidMoment(past = true)
+	@Temporal(TemporalType.TIMESTAMP)
+	private Date				moment;
+
+	@Optional
+	@ValidString(max = 255)
 	@Automapped
-	private String				languageSkills;
+	private String				bio;
+
+	@Optional
+	@ValidMoney
+	@Automapped
+	private Money				salary;
+
+	@Optional
+	@ValidUrl
+	@Automapped
+	private String				photoUrl;
 
 	@Mandatory
 	@Valid
 	@ManyToOne(optional = false)
 	private Airline				airline;
 
-	@Mandatory
-	@Valid
-	@Automapped
-	private AvailabilityStatus	availability;
-
-	@Mandatory
-	@ValidMoney
-	@Automapped
-	private Money				salary;
-
-	@Optional
-	@ValidNumber(integer = 2)
-	@Automapped
-	private Integer				experienceYears;
 }
