+++ src/main/java/acme/realms/flight_crew_members/AvailabilityStatus.java
@@ -1,5 +1,5 @@
 
-package acme.entities.flight_crew_members;
+package acme.realms.flight_crew_members;
 
 public enum AvailabilityStatus {
 	AVAILABLE, ON_VACATION, ON_LEAVE
