+++ src/main/resources/WEB-INF/views/validation-en.i18n
@@ -8,9 +8,8 @@ acme.validation.legs.flight-number.message = The flight number must be composed 
 acme.validation.tracking-log.indicator.message = Error with the percentage and the status of the tracking log.
 acme.validation.tracking-log.resolution.message = The status and the description of the tracking log do not match.
 acme.validation.service.promotionCode.message = The promotion code must follow the net pattern: 4 letters in uppercase , hyphen and 2 digits tha represents the actual year.
-acme.validation.service.promotionCodeAndDiscount.message = Si existe un código de descuento debe existir un descuento y viceversa.
+acme.validation.service.promotionCodeAndDiscount.message = If there is a discount code, there must be a discount and vice versa.
 acme.validation.airline-manager.identifier.message = The indentifier must have 2 or 3 letters correspond to your initials and 6 digits.
 acme.validation.iata-code.message = The IATA code must be 3 letters
-acme.validation.phone-number.message = The phone number must be the character + and must have between 6 and 15 digits.
+acme.validation.phone-number.message = The phone number could be the character + and must have between 6 and 15 digits.
 acme.validation.service.promotionCodeAndDiscount.message = If there is a discount code, there must be a discount and vice versa.
-
